Upgrade GData
- Store entire raw string and build indices for values.
- Test with input mappings, fonts, manifests, etc.
- Pass: Linux, Win64.
- Move building docs from CMake to shell script

+++ src/game/input.cpp
@@ -122,7 +122,7 @@ void input::init(Window& out_mainWindow)
 		}
 	});
 	g_tokens.text = out_mainWindow.registerText([](char c) { g_raw.text.push_back(c); });
-	g_tokens.scroll = out_mainWindow.registerScroll([](f32 x, f32 y) { g_raw.mouseScroll += glm::vec2(x, y); });
+	g_tokens.scroll = out_mainWindow.registerScroll([](f64 x, f64 y) { g_raw.mouseScroll += glm::vec2((f32)x, (f32)y); });
 }
 
 void input::fire()
